Tamas/manufacturer location

+++ src/main/java/ro/sapientia/furniture/service/ManufacturerLocationService.java
@@ -1,8 +1,11 @@
 package ro.sapientia.furniture.service;
 
+import javassist.NotFoundException;
 import org.springframework.stereotype.Service;
+import ro.sapientia.furniture.model.Manufacturer;
 import ro.sapientia.furniture.model.ManufacturerLocation;
 import ro.sapientia.furniture.repository.ManufacturerLocationRepository;
+import ro.sapientia.furniture.util.StatusMessage;
 
 import java.util.List;
 
@@ -18,8 +21,13 @@ public class ManufacturerLocationService {
         return this.manufacturerLocationRepository.findAll();
     }
 
-    public ManufacturerLocation findManufacturerLocationById(final Long id) {
-        return this.manufacturerLocationRepository.findManufacturerLocationById(id);
+    public ManufacturerLocation findManufacturerLocationById(final Long id) throws NotFoundException {
+        ManufacturerLocation location = this.manufacturerLocationRepository.findManufacturerLocationById(id);
+        if (location != null) {
+            return location;
+        } else {
+            throw new NotFoundException(StatusMessage.NOT_FOUND);
+        }
     }
 
     public ManufacturerLocation create(ManufacturerLocation manufacturerLocation) {
@@ -30,7 +38,11 @@ public class ManufacturerLocationService {
         return this.manufacturerLocationRepository.saveAndFlush(manufacturerLocation);
     }
 
-    public void delete(Long id) {
-        this.manufacturerLocationRepository.deleteById(id);
+    public void delete(Long id) throws NotFoundException {
+        if (manufacturerLocationRepository.findManufacturerLocationById(id) != null) {
+            this.manufacturerLocationRepository.deleteById(id);
+        } else {
+            throw new NotFoundException(StatusMessage.NOT_FOUND);
+        }
     }
 }
